must give path to csv file to run the program

+++ run.sh
@@ -5,6 +5,13 @@
 # Download the dependency
 # curl http://www.stuycs.org/courses/apcs/k/2019-01-03/lanterna.jar -o lanterna.ja
 
+# An csv file must be passed in for the program to work.
+if [ $# -eq 0 ]
+then
+    echo "Give the path to the csv file (relative or absolute)
+    exit 1
+fi
+
 rm *.class
 
 # Compile the project
